Remove the weird workarounds and support Result-returning test funs
I hadn't realized that rust supports Results as return values from
test functions; or maybe it didn't when I first wrote this; anyway,
this is what's supported now.

With this change, we support returning more than a single kind of
Result (as long as they all convert into the result that the test
function returns) from the test block, which relied on type inference
before.